+++ lib/legion/cli/chat/tool_registry.rb
@@ -15,6 +15,7 @@ begin
   require 'legion/cli/chat/tools/search_memory'
   require 'legion/cli/chat/tools/web_search'
   require 'legion/cli/chat/tools/spawn_agent'
+  require 'legion/cli/chat/tools/search_traces'
 rescue LoadError => e
   Legion::Logging.debug("ToolRegistry ruby_llm not available, chat tools will not be registered: #{e.message}") if defined?(Legion::Logging)
 end
@@ -36,7 +37,8 @@ module Legion
                             Tools::SaveMemory,
                             Tools::SearchMemory,
                             Tools::WebSearch,
-                            Tools::SpawnAgent
+                            Tools::SpawnAgent,
+                            Tools::SearchTraces
                           ].freeze
                         else
                           [].freeze